Datasets in the STARTS Package

Some datasets for illustration used in the examples of the STARTS package.

data(data.starts01a)
data(data.starts01b)
data(data.starts02)
data(data.starts03a)
data(data.starts03b)
data(data.starts03c)

Format

data.starts01a. A resimulated dataset containing three factors from the Big5 scale measured at five time points used in Luedtke, Robitzsch and Wagner (2018). The dataset only contains observations without missing data.

'data.frame': 890 obs. of 16 variables:
$ id: int 100006 100008 100010 100014 100032 100033 100035 100038 100049 100050 ...
$ E1: num -0.28 1.48 0.12 -1.05 -0.28 ...
$ E2: num 0.12 -0.092 0.495 -0.679 -0.467 ...
$ E3: num 1.08 0.12 0.12 -1.27 -0.28 ...
$ E4: num 0.495 0.12 1.294 -2.229 -0.28 ...
$ E5: num -0.092 0.707 0.707 -2.041 -0.092 ...
$ N1: num 1.114 -0.173 -0.017 0.958 1.27 ...
$ N2: num -0.348 0.003 -1.167 1.602 1.758 ...
$ N3: num -0.192 0.471 -0.348 1.114 0.627 ...
$ N4: num -0.348 -1.167 -0.504 1.426 1.27 ...
$ N5: num -0.192 -0.836 -0.192 2.421 1.27 ...
$ O1: num 1.994 -1.82 -0.107 -0.678 -0.792 ...
$ O2: num 1.423 -0.678 -0.678 -0.678 1.423 ...
$ O3: num 1.423 -1.066 -0.678 0.075 0.852 ...
$ O4: num -0.29 -0.678 -0.29 0.075 -0.107 ...
$ O5: num 1.217 -1.637 -0.29 -0.678 0.646 ...

data.starts01b. Like data.starts01a, but the dataset also contains cases with missing data.

'data.frame': 3215 obs. of 17 variables:
$ id : int 100001 100002 100003 100004 100005 100006 100007 100008 100009 100010 ...
$ patt: Factor w/ 26 levels "P00010","P00011",..: 24 19 20 25 22 26 18 26 19 26 ...
$ E1 : num 0.308 1.67 0.308 0.308 -0.468 ...
$ E2 : num 0.308 0.895 0.707 0.707 0.12 0.12 NA -0.092 -0.28 0.496 ...
$ E3 : num 0.895 NA NA 0.895 NA ...
$ E4 : num NA NA NA 0.496 0.496 ...
$ E5 : num 0.707 NA 0.308 NA 0.496 -0.092 -0.28 0.707 NA 0.707 ...
$ N1 : num 0.783 -0.017 -0.192 -0.017 -0.504 ...
$ N2 : num 1.114 -0.348 -0.348 -0.348 -0.836 ...
$ N3 : num -0.348 NA NA -0.348 NA ...
$ N4 : num NA NA NA -0.504 -1.811 ...
$ N5 : num 0.471 NA -0.192 NA -1.421 ...
$ O1 : num -0.495 -0.107 -0.495 1.035 -0.792 ...
$ O2 : num -0.107 -0.107 -0.29 1.035 -0.29 ...
$ O3 : num 0.464 NA NA 1.423 NA ...
$ O4 : num NA NA NA 1.423 0.281 ...
$ O5 : num 0.646 NA -1.066 NA 0.281 ...

data.starts02 contrains means and covariance matrices of the study of Wu (2016) for the older and the younger cohort (Table 2). Variables a indicate item parcels of negative attitude factor at six occasions. Variable b denotes the performance difficulty factor and variable c the somatic factor.

List of 2
$ older_cohort :List of 3
..$ nobs : num 630
..$ mean : Named num [1:18] 3.53 3.46 3.12 2.71 2.8 2.67 2.62 2.69 2.46 2.37 ...
.. ..- attr(*, "names")=chr [1:18] "a1" "a2" "a3" "a4" ...
..$ covmat:'data.frame': 18 obs. of 18 variables:
$ younger_cohort:List of 3
..$ nobs : num 660
..$ mean : Named num [1:18] 4.62 4.52 4.46 3.58 3.96 3.21 2.94 3.16 3.03 2.74 ...
.. ..- attr(*, "names")=chr [1:18] "a1" "a2" "a3" "a4" ...
..$ covmat:'data.frame': 18 obs. of 18 variables:

data.starts03a contains data from Wagner, Luedtke and Trautwein (2016) of the total sample. data.starts03b contains covariance matrices for both gender groups. data.starts03c contains covariance matrices for both groups of different levels of depression.

The structure of data.starts03a is

List of 2
$ nobs : num 4532
$ covmat: num [1:6, 1:6] 0.236 0.164 0.147 0.129 0.13 ...
..- attr(*, "dimnames")=List of 2
.. ..$ : chr [1:6] "T1" "T2" "T3" "T4" ...
.. ..$ : chr [1:6] "T1" "T2" "T3" "T4" ...

The structure of data.starts03b is

List of 2
$ female:List of 2
..$ nobs : num 2495
..$ covmat: num [1:6, 1:6] 0.22 0.158 0.139 0.18 0.116 ...
.. ..- attr(*, "dimnames")=List of 2
.. .. ..$ : chr [1:6] "T1" "T2" "T3" "T4" ...
.. .. ..$ : chr [1:6] "T1" "T2" "T3" "T4" ...
$ male :List of 2
..$ nobs : num 2037
..$ covmat: num [1:6, 1:6] 0.25 0.165 0.152 0.13 0.147 ...
.. ..- attr(*, "dimnames")=List of 2
.. .. ..$ : chr [1:6] "T1" "T2" "T3" "T4" ...
.. .. ..$ : chr [1:6] "T1" "T2" "T3" "T4" ...

The structure of data.starts03c is

List of 2
$ high:List of 2
..$ nobs : num 1342
..$ covmat: num [1:6, 1:6] 0.24 0.172 0.153 0.191 0.127 ...
.. ..- attr(*, "dimnames")=List of 2
.. .. ..$ : chr [1:6] "T1" "T2" "T3" "T4" ...
.. .. ..$ : chr [1:6] "T1" "T2" "T3" "T4" ...
$ low :List of 2
..$ nobs : num 1742
..$ covmat: num [1:6, 1:6] 0.213 0.12 0.118 0.109 0.12 ...
.. ..- attr(*, "dimnames")=List of 2
.. .. ..$ : chr [1:6] "T1" "T2" "T3" "T4" ...
.. .. ..$ : chr [1:6] "T1" "T2" "T3" "T4" ...

References

Luedtke, O., Robitzsch, A., & Wagner, J. (2018). More stable estimation of the STARTS model: A Bayesian approach using Markov Chain Monte Carlo techniques. Psychological Methods, 23(3), 570-593. doi: 10.1037/met0000155

Wagner, J., Luedtke, O., & Trautwein, U. (2016). Self-esteem is mostly stable across young adulthood: Evidence from latent STARTS models. Journal of Personality, 84(4), 523-535. doi: 10.1111/jopy.12178

Wu, P.-C. (2016). Longitudinal stability of the Beck Depression Inventory II: A latent trait-state-occasion model. Journal of Psychoeducational Assessment, 34, 39-53. doi: 10.1177/0734282915582101